+++ aeromodelo.h
@@ -13,24 +13,24 @@ namespace SimuladorDeVooSimples
     class Aeromodelo : public Aviao
     {
     private:
-        float preco;              // Valor do aeromodelo
+        double preco;              // Valor do aeromodelo
         string cor;               // Cor predominante do aeromodelo
         string modeloTransmissor; // Modelo de transmisor do aeromodelo
 
     public:
         // Construtor que inicializa os parâmetros
-        Aeromodelo(string, string, string, string, float = 0.0, float = 0.0, float = 0.0, float = 0.0, float = 0.0, float = 0.0);
+        Aeromodelo(string, string, string, string, float = 0.0, float = 0.0, float = 0.0, float = 0.0, float = 0.0, double = 0.0);
 
         // Destrutor padrão
         ~Aeromodelo();
 
         // Métodos para mudar as variaveis do aeromodelo
-        void setPreco(float);
+        void setPreco(double);
         void setCor(string);
         void setModeloTransmissor(string);
 
         // Métodos para saber os valores das variáveis do aeromodelo
-        float getPreco();
+        double getPreco();
         string getCor();
         string getModeloTransmissor();
